+++ Sources/Parley/Extensions/ErrorExtension.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension Error {
+    
+    func getFormattedMessage() -> String {
+        guard let parleyResponse = self as? ParleyErrorResponse,
+              let parleyMessage = parleyResponse.notifications.first?.message else {
+            return self.localizedDescription
+        }
+        return parleyMessage
+    }
+}